server/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.15.2
excel-base==1.0.4
isoweek==1.3.3
python-dateutil==2.9.0.post0
psycopg2==2.9.9
psycopg2-binary==2.9.9
screen==1.0.1
six==1.16.0
sqlparse==0.5.0
TimeConvert==3.0.13
typing_extensions==4.12.2
tzlocal==5.2
xlwt==1.3.0

# celery==5.2.7
# gunicorn==20.1.0
# psycopg2-binary==2.9.5
# python-decouple==3.6
# redis==4.3.4
